Move all routes to routes folder for segregation
