+++ src/rest/routes/users/route.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { validate: validateSchema } = require('../middleware/schema');
+const { validate: validateSchema } = require('../../middleware/schema');
 const userSchema = require('./schema.json');
 
 const {
